pyproject.toml
[project]
name = "blog-platform-api"
version = "0.1.0"
description = ""

requires-python = ">=3.12"
dependencies = [
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[crypography] (>=3.5.0,<4.0.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "alembic (>=1.16.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.8"

